[ARQ-1108] - Initial support for client side transaction handling. Includes embedded (local protocol) containers support.
