+++ arquillian-service-integration-spring-3-int-tests/src/test/resources/persistence/persistence.xml
@@ -2,7 +2,8 @@
              xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
              xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd"
              version="2.0">
-    <persistence-unit name="ArquillianTestUnit" transaction-type="RESOURCE_LOCAL">
+
+    <persistence-unit name="ArquillianTestUnit" transaction-type="JTA">
         <provider>org.hibernate.ejb.HibernatePersistence</provider>
         <jta-data-source>jdbc/ArquillianSpringDataSource</jta-data-source>
         <class>org.jboss.arquillian.spring.testsuite.beans.model.Employee</class>
