+++ arquillian-service-integration-spring-3-int-tests/src/test/resources/web.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<web-app xmlns="http://java.sun.com/xml/ns/javaee"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
+         version="3.0">
+
+    <persistence-unit-ref>
+        <persistence-unit-ref-name>persistence/ArquillianTestUnit</persistence-unit-ref-name>
+        <persistence-unit-name>ArquillianTestUnit</persistence-unit-name>
+    </persistence-unit-ref>
+
+</web-app>
