+++ arquillian-transaction-spring/src/main/java/org/jboss/arquillian/transaction/spring/container/ContainerSpringTransactionProvider.java
@@ -1,0 +1,55 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2013, Red Hat Middleware LLC, and individual contributors
+ * by the @authors tag. See the copyright.txt in the distribution for a
+ * full listing of individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.arquillian.transaction.spring.container;
+
+import org.jboss.arquillian.core.api.Instance;
+import org.jboss.arquillian.core.api.annotation.Inject;
+import org.jboss.arquillian.spring.integration.context.RemoteTestScopeApplicationContext;
+import org.jboss.arquillian.spring.integration.context.TestScopeApplicationContext;
+import org.jboss.arquillian.transaction.spring.provider.AbstractSpringTransactionProvider;
+import org.springframework.context.ApplicationContext;
+
+/**
+ * <p>A Spring transaction provider that works with remote container and is able to provide transaction support.</p>
+ *
+ * @author <a href="mailto:jmnarloch@gmail.com">Jakub Narloch</a>
+ * @version $Revision: $
+ */
+public class ContainerSpringTransactionProvider extends AbstractSpringTransactionProvider {
+
+    /**
+     * <p>Instance of application context.</p>
+     */
+    @Inject
+    Instance<RemoteTestScopeApplicationContext> applicationContextInstance;
+
+    /**
+     * <p>Creates new instance of {@link ContainerSpringTransactionProvider}.</p>
+     */
+    public ContainerSpringTransactionProvider() {
+        // empty constructor
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected ApplicationContext getApplicationContext() {
+
+        return applicationContextInstance.get().getApplicationContext();
+    }
+}
