+++ arquillian-transaction-spring/src/test/java/org/jboss/arquillian/transaction/spring/client/SpringTransactionExtensionTestCase.java
@@ -14,30 +14,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.arquillian.spring.integration.transaction.container;
+package org.jboss.arquillian.transaction.spring.client;
 
+import org.jboss.arquillian.container.test.spi.client.deployment.AuxiliaryArchiveAppender;
 import org.jboss.arquillian.core.spi.LoadableExtension;
-import org.jboss.arquillian.transaction.spring.provider.SpringTransactionProvider;
 import org.jboss.arquillian.transaction.spi.provider.TransactionProvider;
-import org.jboss.arquillian.transaction.spring.container.SpringTransactionRemoteExtension;
+import org.jboss.arquillian.transaction.spring.container.ContainerSpringTransactionProvider;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
+import static org.mockito.Mockito.*;
 
 /**
- * <p>Tests {@link org.jboss.arquillian.transaction.spring.container.SpringTransactionRemoteExtension} class.</p>
+ * <p>Tests {@link org.jboss.arquillian.transaction.spring.client.SpringTransactionExtension} class.</p>
  *
  * @author <a href="mailto:jmnarloch@gmail.com">Jakub Narloch</a>
  */
-public class SpringTransactionRemoteExtensionTestCase {
+public class SpringTransactionExtensionTestCase {
 
     /**
      * <p>Represents the instance of tested class.</p>
      */
-    private SpringTransactionRemoteExtension instance;
+    private SpringTransactionExtension instance;
 
     /**
      * <p>Sets up the test environment.</p>
@@ -45,11 +43,11 @@ public class SpringTransactionRemoteExtensionTestCase {
     @Before
     public void setUp() {
 
-        instance = new SpringTransactionRemoteExtension();
+        instance = new SpringTransactionExtension();
     }
 
     /**
-     * <p>Tests the {@link SpringTransactionRemoteExtension#register(LoadableExtension.ExtensionBuilder)} method.</p>
+     * <p>Tests the {@link SpringTransactionExtension#register(LoadableExtension.ExtensionBuilder)} method.</p>
      */
     @Test
     public void testRegister() {
@@ -58,7 +56,8 @@ public class SpringTransactionRemoteExtensionTestCase {
 
         instance.register(mockExtensionBuilder);
 
-        verify(mockExtensionBuilder).service(TransactionProvider.class, SpringTransactionProvider.class);
+        verify(mockExtensionBuilder).service(AuxiliaryArchiveAppender.class, SpringTransactionArchiveAppender.class);
+        verify(mockExtensionBuilder).service(TransactionProvider.class, ContainerSpringTransactionProvider.class);
         verifyNoMoreInteractions(mockExtensionBuilder);
     }
 }
