+++ arquillian-transaction-spring/src/test/java/org/jboss/arquillian/transaction/spring/container/ContainerSpringTransactionProviderTestCase.java
@@ -1,29 +1,12 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2012, Red Hat Middleware LLC, and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.arquillian.spring.integration.transaction.provider;
+package org.jboss.arquillian.transaction.spring.container;
 
 import org.jboss.arquillian.core.api.Instance;
 import org.jboss.arquillian.core.api.InstanceProducer;
 import org.jboss.arquillian.spring.integration.context.RemoteTestScopeApplicationContext;
-import org.jboss.arquillian.spring.integration.context.TestScopeApplicationContext;
-import org.jboss.arquillian.spring.integration.transaction.utils.TestReflectionHelper;
 import org.jboss.arquillian.transaction.api.annotation.Transactional;
 import org.jboss.arquillian.transaction.impl.test.TransactionalTestImpl;
-import org.jboss.arquillian.transaction.spring.provider.SpringTransactionProvider;
+import org.jboss.arquillian.transaction.spring.provider.AbstractSpringTransactionProvider;
+import org.jboss.arquillian.transaction.spring.utils.TestReflectionHelper;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.context.ApplicationContext;
@@ -31,25 +14,23 @@ import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.TransactionDefinition;
 import org.springframework.transaction.TransactionStatus;
 
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.*;
 
 /**
- * <p>Tests {@link org.jboss.arquillian.transaction.spring.provider.SpringTransactionProvider} class.</p>
+ * <p>Tests {@link ContainerSpringTransactionProvider} class.</p>
  *
  * @author <a href="mailto:jmnarloch@gmail.com">Jakub Narloch</a>
  */
-public class SpringTransactionProviderTestCase {
+public class ContainerSpringTransactionProviderTestCase {
 
     /**
      * <p>Represents the instance of tested class.</p>
      */
-    private SpringTransactionProvider instance;
+    private AbstractSpringTransactionProvider instance;
 
     /**
-     * <p>Mock instance of {@link TestScopeApplicationContext}.</p>
+     * <p>Mock instance of {@link RemoteTestScopeApplicationContext}.</p>
      */
     private Instance<RemoteTestScopeApplicationContext> mockTestScopeApplicationContextInstance;
 
@@ -86,7 +67,7 @@ public class SpringTransactionProviderTestCase {
     @Before
     public void setUp() throws Exception {
 
-        instance = new SpringTransactionProvider();
+        instance = new ContainerSpringTransactionProvider();
 
         mockApplicationContext = mock(ApplicationContext.class);
         mockPlatformTransactionManager = mock(PlatformTransactionManager.class);
@@ -104,13 +85,15 @@ public class SpringTransactionProviderTestCase {
         mockPlatformTransactionManagerInstance = mock(InstanceProducer.class);
         mockTransactionStatusInstance = mock(InstanceProducer.class);
 
-        TestReflectionHelper.setFieldValue(instance, "applicationContextInstance", mockTestScopeApplicationContextInstance);
-        TestReflectionHelper.setFieldValue(instance, "transactionManagerInstance", mockPlatformTransactionManagerInstance);
+        TestReflectionHelper.setFieldValue(instance, "applicationContextInstance",
+                mockTestScopeApplicationContextInstance);
+        TestReflectionHelper.setFieldValue(instance, "transactionManagerInstance",
+                mockPlatformTransactionManagerInstance);
         TestReflectionHelper.setFieldValue(instance, "transactionStatusInstance", mockTransactionStatusInstance);
     }
 
     /**
-     * <p>Tests the {@link SpringTransactionProvider#
+     * <p>Tests the {@link AbstractSpringTransactionProvider#
      * beginTransaction(org.jboss.arquillian.transaction.spi.test.TransactionalTest)} method.</p>
      */
     @Test
@@ -124,8 +107,8 @@ public class SpringTransactionProviderTestCase {
     }
 
     /**
-     * <p>Tests the {@link SpringTransactionProvider#
-     * rollbackTransaction(org.jboss.arquillian.transaction.spi.test.TransactionalTest)} method.</p>
+     * <p>Tests the {@link AbstractSpringTransactionProvider#rollbackTransaction(
+     * org.jboss.arquillian.transaction.spi.test.TransactionalTest)} method.</p>
      */
     @Test
     public void testRollbackTransaction() {
@@ -139,8 +122,8 @@ public class SpringTransactionProviderTestCase {
     }
 
     /**
-     * <p>Tests the {@link SpringTransactionProvider#
-     * commitTransaction(org.jboss.arquillian.transaction.spi.test.TransactionalTest)} method.</p>
+     * <p>Tests the {@link AbstractSpringTransactionProvider#
+     * commitTransaction(org.jboss.arquillian.transaction.spi.test.TransactionalTest)}method.</p>
      */
     @Test
     public void testCommitTransaction() {
