+++ pom.xml
@@ -53,7 +53,7 @@
         <version.arquillian_warp>1.0.0.Alpha2</version.arquillian_warp>
 
         <!-- Arquillian Transaction -->
-        <version.arquillian_transaction>1.0.0.Alpha1</version.arquillian_transaction>
+        <version.arquillian_transaction>1.0.0.Alpha3</version.arquillian_transaction>
 
         <!-- Test Dependencies -->
         <version.junit_junit>4.8.1</version.junit_junit>
@@ -202,6 +202,7 @@
                 <artifactId>validation-api</artifactId>
                 <version>1.0.0.GA</version>
             </dependency>
+
             <dependency>
                 <groupId>org.hibernate</groupId>
                 <artifactId>hibernate-validator</artifactId>
